+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Support for `@annually` and `@midnight`
+
 ### Changed
 - Function order in Quantum.Matcher
 - Renamed private translate function to do_translate
@@ -76,7 +79,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Initial commit
 
 
-[unreleased]: https://github.com/c-rack/quantum-elixir/compare/v1.2.0...HEAD
+[unreleased]: https://github.com/c-rack/quantum-elixir/compare/v1.2.1...HEAD
 [1.2.1]: https://github.com/c-rack/quantum-elixir/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/c-rack/quantum-elixir/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/c-rack/quantum-elixir/compare/v1.0.4...v1.1.0